refactor(cms): lissage des noms de fichier de usecase
