+++ src/server/cms/domain/actualité.ts
@@ -1,7 +1,7 @@
 import { Article } from './article';
 import { Image } from './image';
 
-export interface Actualite {
+export interface Actualité {
 	article?: Article
 	bannière?: Image
 	extraitContenu: string
