+++ src/server/cms/infra/repositories/strapi.mapper.ts
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 
 import { Image as ImageProps } from '~/client/components/props';
-import { Actualite } from '~/server/cms/domain/actualite';
+import { Actualité } from '~/server/cms/domain/actualité';
 import { AnnonceDeLogement } from '~/server/cms/domain/annonceDeLogement.type';
 import { Article } from '~/server/cms/domain/article';
 import { Question } from '~/server/cms/domain/FAQ.type';
@@ -89,11 +89,11 @@ function mapCartesMesuresEmployeursList(strapiLesMesuresEmployeursDispositif: St
 	};
 }
 
-export function mapStrapiListeActualités(strapiListeActualités: Strapi.SingleType.ListeActualités): Actualite[] {
+export function mapStrapiListeActualités(strapiListeActualités: Strapi.SingleType.ListeActualités): Actualité[] {
 	return strapiListeActualités.listeActualites.map(mapStrapiActualité);
 }
 
-function mapStrapiActualité(strapiActualité: Strapi.SingleType.ListeActualités.Actualité): Actualite {
+function mapStrapiActualité(strapiActualité: Strapi.SingleType.ListeActualités.Actualité): Actualité {
 	const article = flatMapSingleRelation(strapiActualité.article, mapArticle);
 	return {
 		article,
