+++ src/server/cms/useCases/récupérerActualités.useCase.ts
@@ -1,11 +1,11 @@
-import { Actualite } from '~/server/cms/domain/actualite';
+import { Actualité } from '~/server/cms/domain/actualité';
 import { CmsRepository } from '~/server/cms/domain/cms.repository';
 import { Either } from '~/server/errors/either';
 
-export class RécupererActualitesUseCase {
+export class RécupérerActualitésUseCase {
 	constructor(private cmsRepository: CmsRepository) {}
 
-	async handle(): Promise<Either<Actualite[]>> {
+	async handle(): Promise<Either<Actualité[]>> {
 		return this.cmsRepository.getActualitéList();
 	}
 }
